+++ Tests/Objects/IGListTestSection.h
@@ -15,9 +15,8 @@
 @interface IGListTestSection : IGListSectionController
 
 @property (nonatomic, assign) NSInteger items;
-
 @property (nonatomic, assign) CGSize size;
-
 @property (nonatomic, assign) BOOL wasSelected;
+@property (nonatomic, assign) BOOL wasDeselected;
 
 @end